+++ kernel/CHANGE_LOG.md
@@ -1,2 +1,12 @@
+------------------------------------------------------------------------------------------------------------------------
+Date: 2025.07.31
+Version: v0.1.0-alpha.2
+Changes:
+- removed event: EVENT_APP_BEFORE_RUN
+- added function: app.RegisterComponent()
+- refactored ITemplateRenderer
+- changed event payload: EVENT_APP_BEFORE_HANDLE_REQUEST "resource" IHttpResource -> "context" IHandleContext
+
+------------------------------------------------------------------------------------------------------------------------
 Date: 2025.07.24
 Version: v0.1.0-alpha.1